Moved interfaz1 dirs to separate config file
Most of the directories in interfaz1 where absolute paths, which
generated problems when executing from different computers.

In this commit, we changed all the addresses for the relative format
'../' which allows us to use the same dirs for everyone. Also, we
moved all the dir strings to a separate YAML file, which allows
further configuration in an easier way.