tentatively placed the buttons proc & add face
The intent is to have the buttons process and add face working in the
interface. So they were set up to send their output to the textbox,
while getting a directory query as an input.

additional modifications include changes to the scripting backend,
making sure all the relevant variables are accounted for in the
config file and no hardcoding remains.