fixed add_face and evaluate_face buttons
Finally, it was possible to make sure all buttons in the program
work! the main issue with adding faces was a stupid directory
format thing. Given that the directories will just be handed over
to the Docker functions, they must be in docker-absolute format.

then there was a problem with the previously working evaluate-face
function. It was determined that for some reason, the output text
in docker now included two TensorFlow-related lines. It was simply
a matter of changing the selected line at the parsing part of the
procedure.