fixed labels for retrain
This simple commit adds a bunch of shit and fixes one important
but completely ignored thing till now. I'm talking about the
labels which are created in relation to the current dataset.

The following fix is manual, meaning the user needs to keep track
of the number of classes currently supported.